requirements.txt
channels==3.0.4
Django==3.2.10
djangorestframework==3.13.1
django-cors-headers==3.10.1
hypercorn==0.13.2
django_heroku
daphne
whitenoise
psycopg2
dj-database-url